not gonna touch the code

+++ 9-binary_tree_height.c
@@ -26,7 +26,7 @@ size_t binary_tree_height(const binary_tree_t *tree)
  * Return: max value between a and b
  */
 
-int max(int a, int b)
+size_t max(size_t a, size_t b)
 {
 	if (a >= b)
 		return (a);
